+++ src/ekf_imu.py
@@ -10,6 +10,7 @@ from utils.misc import (
     plot_error,
 )
 from utils.models import (
+    Measurement,
     AltitudeById,
 )
 from utils.imu import ( 
@@ -22,9 +23,9 @@ from utils.states import (
     CompositeState, 
     StateWithCovariance,
 )
-from utils.meas import (
-    Measurement,
+from utils.misc import (
     RangeData,
+    height_measurements,
 )
 from miluv.data import DataLoader
 from src.filters import ExtendedKalmanFilter
@@ -52,7 +53,7 @@ save_fig = True
 save_results = False
 
 
-""" Get data """
+""" Load data """
 # parser = argparse.ArgumentParser()
 # parser.add_argument('--exp', required=True)
 # args = parser.parse_args()
@@ -61,81 +62,47 @@ exp = "1c"
 folder = "/media/syedshabbir/Seagate B/data"
 miluv = DataLoader(exp, exp_dir = folder, barometer = False)
 
-# Preliminaries
+""" Preliminaries """
 robots = list(miluv.data.keys())
 input_sensor = "imu_px4"
 input_freq = 190
-
-# Get the time range
-start_time, end_time = miluv.get_timerange(
-                      sensors = input_sensor)
+start_time, end_time = miluv.get_timerange(sensors = input_sensor)
 end_time = start_time +2
-query_stamps = np.arange(start_time, 
-                         end_time, 
-                         1/input_freq)
+query_stamps = np.arange(start_time, end_time, 1/input_freq)
 
 """ Get Data """
-imus = miluv.by_timestamps(query_stamps, 
-                           sensors=input_sensor)
+imus = miluv.by_timestamps(query_stamps, sensors=input_sensor)
+pose = [miluv.data[robot]['mocap'].SE23_state(query_stamps) for robot in robots]
+range_data = miluv.by_timerange(start_time, end_time, sensors=['uwb_range'])
 
-height = miluv.by_timerange(start_time, 
-                            end_time, 
-                            sensors=["height"])
-
-height = [height.data[robot]["height"] for robot in robots]
-min_height = [h['range'].min() for h in height]
-
-pose = [miluv.data[robot]["mocap"].extended_pose_matrix(
-        query_stamps) for robot in robots]
-
-range_data = miluv.by_timerange(start_time, end_time,
-                                sensors=["uwb_range"])
+""" Create measurements """
 range_data = RangeData(range_data)
 range_data = range_data.filter_by_bias( max_bias=0.3)
-""" Set Data """
-
-# Range measurements
-meas_data = range_data.to_measurements(
-    reference_id = 'world')
-
-# Height measurements
+meas_data = range_data.to_measurements(reference_id = 'world')
 R = [3*0.1, 3*0.1, 3*0.1]
 bias = [-0.0924, -0.0088, -0.1207]
-for n, robot in enumerate(robots):
-    for i in range(len(height[n])):
-        y = Measurement(value = height[n].iloc[i]['range'],
-                        stamp = height[n].iloc[i]['timestamp'],
-                        model = AltitudeById(R = R[n], 
-                        state_id = robot,
-                        minimum=min_height[n],
-                        bias = bias[n]))
-        meas_data.append(y)
-
-# sort the measurements
+height = height_measurements(miluv, start_time, end_time,  R, bias)
 meas_data = sorted(meas_data, key=lambda x: x.stamp)
 
 """ Create ground truth data """
 ground_truth = []
-bias_gyro = [0.0, 0.0, 0.0]
-bias_accel = [0.0, 0.0, 0.0]
 for i in range(len(query_stamps)):
     x = [IMUState(  SE23_state = pose[n][i], 
-                    bias_gyro = bias_gyro,
-                    bias_accel = bias_accel,
+                    bias_gyro = [0,0,0],
+                    bias_accel =[0,0,0],
                     state_id = robot,
                     stamp = query_stamps[i],) 
-        for n, robot in enumerate(robots)
-        ]
-    ground_truth.append(CompositeState(x,
-                        stamp = query_stamps[i]))
+        for n, robot in enumerate(robots) ]
+    ground_truth.append(CompositeState(x, stamp = query_stamps[i]))
     
 """ State and input covariance, process model, and initial state"""
+
 # Initial state
 x0 = ground_truth[0].copy()
 
 # State and input covariance
 P0 = np.diag([0.1, 0.1, 0.1, 
-             0.1, 0.1, 0.1,
+                0.1, 0.1, 0.1,
              0.01, 0.01, 0.1,
              0.0001, 0.0001,0.0001,
              0.0001, 0.0001,0.0001])
@@ -158,14 +125,14 @@ for n, robot in enumerate(robots):
     dt = 1/input_freq
     Q[robot] = Q_c / dt
 
-process_model = CompositeProcessModel(
-    [IMUKinematics(Q[robot]) for robot in robots])
+process_model = CompositeProcessModel([IMUKinematics(Q[robot]) for robot in robots])
 
 # Composite covariance
 n_states = len(x0.value)
 P0 = np.kron(np.eye(n_states), P0)
 
 """ Create input data """
+
 input_data = []
 for i in range(len(query_stamps)):
     u = [Input(
@@ -179,54 +146,40 @@ for i in range(len(query_stamps)):
              'linear_acceleration.z']].values,
         stamp = query_stamps[i], 
         state_id = robot)
-        for robot in robots
-        ]
-    
+        for robot in robots]
     input_data.append(u)
+
+""" Implement your filter here """
 
 if ekf:
     """ Run the filter """
     x = StateWithCovariance(x0, P0)
-
-    # Try an EKF or an IterEKF
-    ekf = ExtendedKalmanFilter(process_model, 
-                            reject_outliers=True)
+    ekf = ExtendedKalmanFilter(process_model, reject_outliers=True)
 
     meas_idx = 0
     y = meas_data[meas_idx]
     init_stamp = time.time()
     results_list = []
     for k in tqdm(range(len(input_data) - 1)):
-
         u = input_data[k]
                 
         # Fuse any measurements that have occurred.
         while y.stamp < input_data[k + 1][0].stamp and meas_idx < len(meas_data):
-
             x = ekf.correct(x, y, u)
-
             # Load the next measurement
             meas_idx += 1
             if meas_idx < len(meas_data):
                 y = meas_data[meas_idx]
-
         dt = input_data[k + 1][0].stamp - x.stamp
         x = ekf.predict(x, u, dt)
 
         results_list.append(GaussianResult(x, ground_truth[k]))
-
-    print("Average filter computation frequency (Hz):")
-    print(1 / ((time.time() - init_stamp) / len(input_data)))
-
     results = GaussianResultList(results_list)
 
-""" Print position RMSE """
-script_dir = os.path.dirname(
-    os.path.abspath(sys.argv[0]))
-
+script_dir = os.path.dirname(os.path.abspath(sys.argv[0]))
 if ekf:
+    """ Print position RMSE """
     pos_rmse = {robot: {} for robot in robots}
-    dof = 3
     for i, robot in enumerate(robots):
         pos = np.array([r.get_state_by_id(robot).value[0].value[0:3, -1]
                         for r in results.state])
@@ -239,6 +192,7 @@ if ekf:
         print(f"Position RMSE for Experiment: {exp} and robot {robot}: {pos_rmse[robot]} m")
 
 if ekf and save_results:
+    """ Save results into pkl file """
     folder = os.path.join(script_dir, f'results')
     os.umask(0)
     os.makedirs(folder, exist_ok=True)
@@ -248,6 +202,7 @@ if ekf and save_results:
         pickle.dump((results, pos_rmse), file)
 
 if ekf and error_plot:
+    """ Plot error """
     separate_figs = True
     figs = plot_error(results, 
                       separate_figs=separate_figs)
