+++ utils/__init__.py
@@ -1,3 +1,4 @@
 from .imu import *
+from .models import *
 from .states import *
-from .meas import *
+from .misc import *
